+++ fcnui-website/docs/CHANGELOG.md
@@ -13,6 +13,10 @@ Full changelog of the project.
 - [CLI] Remove `get_it` and `equatable` dependencies.
 - [CLI] Added some tests. (Not complete).
 - [Base] Add new dependencies to README.md.
+- [Registry] Add `Checkbox` component.
+- [Playground] Add `checkbox` page.
+- [Registry] Add `PageImpl` that makes pages for Playground.
+- [Homepage] Add `checkbox` documentation.
 
 ## 18.02.2024